xrpledger/requirements.txt
anyio==4.8.0
base58==2.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptoconditions==0.8.1
cryptography==3.4.7
Deprecated==1.2.18
ECPy==1.2.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pyasn1==0.4.8
pycparser==2.22
pycryptodome==3.21.0
PyNaCl==1.4.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
toml==0.10.2
types-Deprecated==1.2.15.20241117
typing_extensions==4.12.2
urllib3==2.3.0
websockets==13.1
wrapt==1.17.2
wxPython==4.2.2
xrpl-py==4.0.0
